Safari-611.1.10/LayoutTests/animations/animation-events-create-expected.txt

Test dynamic creation of WebKitAnimationEvent.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.createEvent exists

Test creation of WebKitAnimationEvent
PASS typeof(ev) is 'object'
PASS ev.animationName is ''
PASS ev.elapsedTime is 0.0

Create a MouseEvent and make sure it doesn't have the WebKitAnimationEnd event properties
PASS typeof(ev) is 'object'
PASS ev.animationName is undefined
PASS ev.elapsedTime is undefined

PASS successfullyParsed is true

TEST COMPLETE
